Handle deletion of tags and categories

+++ backend/app/Http/Controllers/TagsController.php
@@ -94,6 +94,13 @@ class TagsController extends Controller
      */
     public function destroy(Tag $tag)
     {
+        if ($tag->posts->count() > 0) {
+            session()->flash(
+                'error',
+                'Tag cannot be deleted because it is associated to ' . $tag->posts->count() . ' posts'
+            );
+            return redirect()->back();
+        }
         $tag->delete();
 
         session()->flash('success', 'Tag deleted successfully');
